refactor: improve random number generation and variable initialization

+++ Tienda_Equipos_Audio/CreadorSistemaPersonalizado.cpp
@@ -5,7 +5,7 @@ Componente* CreadorSistemaPersonalizado::crearSistemaDeAudio()
     // Un sistema de audio tiene un procesador de senal, una fuente de audio y un parlante
 
     // el codigo debe de ser una cadena string de 5 numeros aleatorios
-    srand(time(NULL));
+    srand(static_cast<unsigned int>(time(NULL)));   // genera una semilla para el generador de numeros aleatorios (se usa static_cast para evitar un warning)
     std::string codigo = "";
     for (int i = 0; i < 5; i++)
     {
